+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 -   Adds the `sort_by` filter to sort lists of posts by property. Added sorting by date. Fall back `.date` and `.updated` to file attributes. ([#33](https://github.com/marceljs/marcel/issues/33))
 -   Posts that have `draft: true` in their front-matter will not be built / served by default. Pass in `-d` or `--drafts` to the CLI commands to include the draft posts. ([#41](https://github.com/marceljs/marcel/issues/41)) Posts that have `permalink: false` in their front-matter will not be included in the build.
 -   Posts that have their `slug` / `permalink` ending in `.html` will not get `/index.html` appended to them.
+-   Adds `post.template` to be able to add a specific template to the post.
 
 ### 0.1.3